buf fix (server) - 20230302
- 플레이어가 방을 나갈 시 방에는 계속 존재한다고 뜨는 부분 수정

+++ Graduation_Server/Graduation_Server/protocol.h
@@ -56,8 +56,8 @@ struct Position {
 };
 
 struct Roominfo_by10 {
-	int						room_number;
-	int						join_member;
+	short						room_number;
+	short						join_member;
 	GAME_ROOM_STATE::TYPE	state;
 };
